choice selection and VSIDS improved, with the help of kissat
the VSIDS implementation now follows, mostly, the implementation used in kissat

in particular, an exponent is used to increase the VSIDS addition proportional to the number of conflicts
a benefit of this is that there's no need to apply decay to every variable
though, this requires rescoring all the variables on occassion
the kissat source was consulted for details on how this works

a custom heap structure has been introduced with much inspiration from the kissat source
roughly, it's a heap for items which have a fixed index, and the current position of an item is tracked
at the moment this tracking isn't put to use, as there's nothing in the heap other than a pointer to a variable
still this should be useful in the future

to help make prototyping some things easier literals now store a pointer to the variable they wrap
to help indicate intent, etc. variables are pinned, etc.

the implementation of the heap is generic over anything which implements order and the 'fixed index' trait, intend to be implemented by items which have a fixed index in a array/vec etc. for direct access

the heap implementation requires refining and is incomplete, but works ok enough for a commit

with these updates bmc 12 is much improved --- something like ~8x as fast as prior to the lazy structure introduction

+++ src/otter_lib/config/mod.rs
@@ -1,7 +1,7 @@
 pub mod defaults;
 
-pub type ActivityConflict = f32;
-pub type DecayFactor = f32;
+pub type ActivityConflict = f64;
+pub type DecayFactor = f64;
 pub type DecayFrequency = usize;
 pub type GlueStrength = u32;
 pub type LubyConstant = usize;
